requirements.txt
Flask
yt-dlp
googletrans==4.0.0-rc1
langdetect
deep-translator
google-genai
ffmpeg-python